fix the typing on the ID fields on the models and add an endpoint to update a team

+++ src/main/java/org/bnbalint/snackdaddy/models/SuggestedAllergy.java
@@ -18,7 +18,7 @@ public class SuggestedAllergy {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(insertable = false, updatable = false)
-    private long id;
+    private Long id;
 
     @Column
     private String name;
@@ -44,11 +44,11 @@ public class SuggestedAllergy {
     //-------------------------------------------
     // Getters and Setters
     //
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
